refactor(frontend): move splitLines to src/ top level
Drop the utils/ directory and place lines.ts as a flat module under
src/, matching the existing convention used by sourcePanel.ts.
