+++ internal/frontend/src/lines.test.ts
