+++ internal/frontend/src/lines.ts
